Rename directory and validator
