+++ modules/appeals_api/config/locales/en.yml
@@ -6,8 +6,10 @@ en:
         hearing_type_preference_missing: "If '/data/attributes/boardReviewOption' 'hearing' is selected, '/data/attributes/hearingTypePreference' must also be present"
         nod_not_found: "NoticeOfDisagreement with uuid %{id} not found."
         not_homeless_address_missing: "If not homeless, address must be provided: '/data/attributes/veteran/address'"
-      uploads:
+      evidence_submission:
         max_dimensions_error: "File exceeds the maximum page dimensions of %{max_dimensions}"
         max_file_size_error: "File cannot exceed a file size of %{max_file_size}"
+        pdf_extension_message: ".pdf is likely an incorrect extension for this document"
+        pdf_read_error: "PdfInfo::MetadataReadError: Syntax Warning: May not be a PDF file"
         successful_validation: "File validated"
         unsupported_file_type: "File must be in PDF format"
